requirements.txt
httpx>=0.27.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
Pillow>=10.4.0
